fix: compoennts collision naming
