+++ frontend/components/avatar/AvatarEmpty.tsx
@@ -1,6 +1,19 @@
 import { HTMLAttributes } from "react";
 import { cva, VariantProps } from "cva";
 
+interface AvatarEmptyProps extends HTMLAttributes<HTMLDivElement>, VariantProps<typeof avatarEmptyCVA> {
+  title: string;
+}
+
+export default function AvatarEmpty(props: AvatarEmptyProps) {
+  const { className, title, figure, size, border, background, ...additionalProps } = props;
+  return (
+    <div className={avatarEmptyCVA({ figure, size, border, background, className })} {...additionalProps}>
+      <p>{title[0]?.toUpperCase()}</p>
+    </div>
+  );
+}
+
 const avatarEmptyCVA = cva(["flex", "justify-center", "items-center", "font-bold"], {
   variants: {
     figure: {
@@ -45,16 +58,3 @@ const avatarEmptyCVA = cva(["flex", "justify-center", "items-center", "font-bold
     background: "gray",
   },
 });
-
-interface AvatarEmptyProps extends HTMLAttributes<HTMLDivElement>, VariantProps<typeof avatarEmptyCVA> {
-  title: string;
-}
-
-export default function AvatarEmpty(props: AvatarEmptyProps) {
-  const { className, title, figure, size, border, background, ...additionalProps } = props;
-  return (
-    <div className={avatarEmptyCVA({ figure, size, border, background, className })} {...additionalProps}>
-      <p>{title[0]?.toUpperCase()}</p>
-    </div>
-  );
-}
