+++ frontend/components/avatar/BasicAvatar.tsx
@@ -2,6 +2,27 @@ import { cva, VariantProps } from "cva";
 import { ImgHTMLAttributes } from "react";
 import AvatarEmpty from "./AvatarEmpty";
 
+interface AvatarProps extends ImgHTMLAttributes<HTMLImageElement>, VariantProps<typeof avatarCVA> {
+  title: string;
+}
+
+export default function Avatar(props: AvatarProps) {
+  const { figure, size, border, className, src, title, ...additionalProps } = props;
+
+  if (!src) {
+    return <AvatarEmpty figure={figure} size={size} border={border} className={className} title={title} />;
+  }
+
+  return (
+    <img
+      src={src}
+      alt="Internet Computer"
+      className={avatarCVA({ figure, size, border, className })}
+      {...additionalProps}
+    />
+  );
+}
+
 const avatarCVA = cva([], {
   variants: {
     figure: {
@@ -34,24 +55,3 @@ const avatarCVA = cva([], {
     size: "medium",
   },
 });
-
-interface AvatarProps extends ImgHTMLAttributes<HTMLImageElement>, VariantProps<typeof avatarCVA> {
-  title: string;
-}
-
-export default function Avatar(props: AvatarProps) {
-  const { figure, size, border, className, src, title, ...additionalProps } = props;
-
-  if (!src) {
-    return <AvatarEmpty figure={figure} size={size} border={border} className={className} title={title} />;
-  }
-
-  return (
-    <img
-      src={src}
-      alt="Internet Computer"
-      className={avatarCVA({ figure, size, border, className })}
-      {...additionalProps}
-    />
-  );
-}
