+++ frontend/components/button/CustomButton.tsx
@@ -1,6 +1,12 @@
 import { ButtonHTMLAttributes } from "react";
 import { cva, type VariantProps } from "cva";
 
+export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof button> {}
+
+export default function CustomButton({ className, intent, size, border, focusBorder, ...props }: ButtonProps) {
+  return <button className={button({ intent, size, border, focusBorder, className })} {...props} />;
+}
+
 const button = cva("button", {
   variants: {
     intent: {
@@ -40,9 +46,3 @@ const button = cva("button", {
     focusBorder: "none",
   },
 });
-
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement>, VariantProps<typeof button> {}
-
-export default function CustomButton({ className, intent, size, border, focusBorder, ...props }: ButtonProps) {
-  return <button className={button({ intent, size, border, focusBorder, className })} {...props} />;
-}
