+++ frontend/components/checkbox/index.ts
@@ -1,2 +1,2 @@
-export { default as CheckBox } from "./CheckBox";
+export { default as BasicCheckBox } from "./BasicCheckBox";
 export { default as CustomCheck } from "./CustomCheck";
