+++ frontend/components/chip/index.ts
@@ -1,1 +1,1 @@
-export { default as Chip } from "./Chip";
+export { default as BasicChip } from "./BasicChip";
