+++ frontend/components/drawer/BasicDrawer.tsx
@@ -1,5 +1,4 @@
 import { ReactComponent as CloseIcon } from "@assets/svg/files/close.svg";
-import { getDrawerBlank, getDrawerContainerStyle } from "./styles.cva";
 import { clsx } from "clsx";
 
 interface IDrawerProps {
@@ -34,5 +33,40 @@ function getCloseIconStyles(enabled: boolean) {
     enabled
       ? "stroke-PrimaryTextColorLight dark:stroke-PrimaryTextColor"
       : "stroke-PrimaryTextColorLight/50 dark:stroke-PrimaryTextColor/50 cursor-not-allowed",
+  );
+}
+
+function getDrawerContainerStyle(isDrawerOpen: boolean) {
+  return clsx(
+    "fixed",
+    "px-8 py-4",
+    "z-[1000]",
+    "w-[28rem]",
+    "h-screen",
+    "top-0",
+    "bg-PrimaryColorLight",
+    "dark:bg-PrimaryColor",
+    "transition-{right}",
+    "duration-500",
+    "ease-in-out",
+    "flex flex-col",
+    isDrawerOpen ? "right-0" : "-right-[28rem]",
+  );
+}
+
+function getDrawerBlank(isDrawerOpen: boolean) {
+  return clsx(
+    "fixed",
+    "top-0",
+    "bottom-0",
+    "left-0",
+    "right-0",
+    "z-50",
+    "bg-black",
+    "opacity-50",
+    "transition-opacity",
+    "duration-500",
+    "ease-in-out",
+    !isDrawerOpen && "hidden",
   );
 }
