+++ frontend/components/drawer/index.tsx
@@ -1,1 +1,1 @@
-export {default as Drawer} from "./Drawer";
+export {default as BasicDrawer} from "./BasicDrawer";
