+++ frontend/components/input/index.tsx
@@ -1,3 +1,3 @@
 export { default as CurrencyInput } from "./CurrencyInput";
-export { default as Input } from "./Input";
+export { default as BasicInput } from "./BasicInput";
 export { default as CustomInput } from "./CustomInput";
