+++ frontend/components/modal/index.tsx
@@ -1,1 +1,1 @@
-export { default as Modal } from "./Modal";
+export { default as BasicModal } from "./BasicModal";
