+++ frontend/components/tooltip/index.ts
@@ -1,2 +1,2 @@
-export { default as ToolTip } from "./ToolTip";
+export { default as BasicToolTip } from "./BasicToolTip";
 export { default as CustomCopy } from "./CustomCopy";
