+++ frontend/pages/components/topbar/index.tsx
@@ -21,7 +21,7 @@ import { clsx } from "clsx";
 import * as DropdownMenu from "@radix-ui/react-dropdown-menu";
 import { AccountHook } from "@hooks/accountHook";
 import { logout } from "@redux/CheckAuth";
-import {Modal} from "@components/modal";
+import {BasicModal} from "@components/modal";
 import ThemeModal from "./themeModal";
 import { ThemesEnum } from "@/const";
 import { CustomCopy } from "@components/tooltip";
@@ -161,12 +161,12 @@ const TopBarComponent = ({ isLoginPage }: { isLoginPage: boolean }) => {
           </DropdownMenu.Root>
         </div>
       </div>
-      <Modal open={themeOpen} top="top-[35%]">
+      <BasicModal open={themeOpen} top="top-[35%]">
         <ThemeModal setOpen={setThemeOpen} />
-      </Modal>
-      <Modal open={dbLocationOpen} top="top-[35%]">
+      </BasicModal>
+      <BasicModal open={dbLocationOpen} top="top-[35%]">
         <DbLocationModal setOpen={setDbLocationOpen} />
-      </Modal>
+      </BasicModal>
     </Fragment>
   );
