+++ frontend/pages/contacts/index.tsx
@@ -1,7 +1,7 @@
 import Menu from "@pages/components/Menu";
 import useContactFilters from "./hooks/useContactFilters";
 import ContactFilters from "./components/contactFilters";
-import {Modal} from "@components/modal";
+import {BasicModal} from "@components/modal";
 import AddContact from "./components/ICRC/AddContact";
 import ContactList from "./components/contactList";
 
@@ -25,14 +25,14 @@ const Contacts = () => {
         />
         <ContactList searchKey={searchKey} assetFilter={assetFilter} />
       </div>
-      <Modal
+      <BasicModal
         open={addOpen}
         width="w-[48rem]"
         padding="py-5 px-8"
         border="border border-BorderColorTwoLight dark:border-BorderColorTwo"
       >
         <AddContact setAddOpen={setAddOpen} />
-      </Modal>
+      </BasicModal>
     </div>
   );
 };
