+++ frontend/pages/home/components/ICRC/allowance/AddAllowanceDrawer.tsx
@@ -1,5 +1,5 @@
 import { t } from "i18next";
-import { Drawer } from "@components/drawer";
+import { BasicDrawer } from "@components/drawer";
 import useAllowanceDrawer from "@pages/home/hooks/useAllowanceDrawer";
 import CreateForm from "./CreateForm";
 
@@ -7,13 +7,13 @@ export default function AddAllowanceDrawer() {
   const { isCreateAllowance, onCloseCreateAllowanceDrawer, isLoading } = useAllowanceDrawer();
 
   return (
-    <Drawer
+    <BasicDrawer
       isDrawerOpen={isCreateAllowance}
       onClose={onCloseCreateAllowanceDrawer}
       title={`${t("allowance.add.allowance")}`}
       enableClose={!isLoading}
     >
       {!isCreateAllowance ? null : <CreateForm />}
-    </Drawer>
+    </BasicDrawer>
   );
 }
