+++ frontend/pages/home/components/ICRC/allowance/UpdateAllowanceDrawer.tsx
@@ -1,4 +1,4 @@
-import { Drawer } from "@components/drawer";
+import { BasicDrawer } from "@components/drawer";
 import useAllowanceDrawer from "@pages/home/hooks/useAllowanceDrawer";
 import { useTranslation } from "react-i18next";
 import UpdateForm from "./UpdateForm";
@@ -8,13 +8,13 @@ export default function UpdateAllowanceDrawer() {
   const { isUpdateAllowance, onCloseUpdateAllowanceDrawer, isLoading } = useAllowanceDrawer();
 
   return (
-    <Drawer
+    <BasicDrawer
       isDrawerOpen={isUpdateAllowance}
       onClose={onCloseUpdateAllowanceDrawer}
       title={t("allowance.edit.allowance")}
       enableClose={!isLoading}
     >
       {isUpdateAllowance ? <UpdateForm /> : <></>}
-    </Drawer>
+    </BasicDrawer>
   );
 }
