+++ frontend/pages/home/components/ICRC/detail/transaction/DialogSendConfirmation.tsx
@@ -2,7 +2,7 @@
 import { ReactComponent as CloseIcon } from "@assets/svg/files/close.svg";
 import UpAmountIcon from "@assets/svg/files/up-amount-icon.svg";
 //
-import {Modal} from "@components/modal";
+import {BasicModal} from "@components/modal";
 import { CustomCopy } from "@components/tooltip";
 import { shortAddress } from "@/utils";
 import { ProtocolType, SendingStatusEnum } from "@/const";
@@ -25,7 +25,7 @@ const DialogSendConfirmation = ({ setDrawerOpen, showConfirmationModal, modal }:
   const { t } = useTranslation();
 
   return (
-    <Modal
+    <BasicModal
       open={modal}
       width="w-[22rem]"
       padding="py-3 px-1"
@@ -72,7 +72,7 @@ const DialogSendConfirmation = ({ setDrawerOpen, showConfirmationModal, modal }:
           </div>
         </div>
       </div>
-    </Modal>
+    </BasicModal>
   );
 
   function onClose() {
